+++ src/index.html
@@ -19,7 +19,7 @@
           <input type="text" class="form-control input" placeholder="Add to your list...">
           <input type="submit" class="sub">
         </form>
-        <ul class="list-data">
+        <ul class="list-data d-flex">
         </ul>
       </div>
       <div class="d-flex p-3">
